Rename package from *.pact.spargel to *.spargel.java
