--- stratosphere-addons/spargel/src/main/java/eu/stratosphere/pact/spargel/examples/connectedcomponents/DuplicateLongInputFormat.java
@@ -1,41 +1,0 @@
-/***********************************************************************************************************************
- *
- * Copyright (C) 2010 by the Stratosphere project (http://stratosphere.eu)
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- *
- **********************************************************************************************************************/
-
-package eu.stratosphere.pact.spargel.examples.connectedcomponents;
-
-import eu.stratosphere.pact.common.io.TextInputFormat;
-import eu.stratosphere.pact.common.type.PactRecord;
-import eu.stratosphere.pact.common.type.base.PactLong;
-import eu.stratosphere.pact.common.type.base.parser.DecimalTextLongParser;
-
-public class DuplicateLongInputFormat extends TextInputFormat {
-	
-	private static final long serialVersionUID = 1L;
-	
-	private final PactLong l1 = new PactLong();
-	private final PactLong l2 = new PactLong();
-	
-	@Override
-	public boolean readRecord(PactRecord target, byte[] bytes, int offset, int numBytes) {
-		final long value = DecimalTextLongParser.parseField(bytes, offset, numBytes, (char) 0xffff);
-
-		this.l1.setValue(value);
-		this.l2.setValue(value);
-		
-		target.setField(0, this.l1);
-		target.setField(1, this.l2);
-		return true;
-	}
-}
