+++ stratosphere-addons/spargel/src/main/java/eu/stratosphere/spargel/java/VertexUpdateFunction.java
@@ -12,7 +12,7 @@
  * specific language governing permissions and limitations under the License.
  *
  **********************************************************************************************************************/
-package eu.stratosphere.pact.spargel;
+package eu.stratosphere.spargel.java;
 
 import java.io.Serializable;
 import java.util.Iterator;
